scripts/requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
